+++ assets/js/app.js
@@ -10,6 +10,7 @@ import ResendVerificationEmailPage from "@js/pages/ResendVerificationEmailPage.v
 import ProfilePage from "@js/pages/ProfilePage.vue";
 import WishItemMinePage from "@js/pages/WishItemMinePage.vue";
 import WishItemSearchPage from "@js/pages/WishItemSearchPage.vue";
+import Toast from "@js/components/Toast.vue";
 
 const pages = {
     landing: LandingPage,
@@ -34,5 +35,23 @@ if (appElement) {
         const app = createApp(pages[pageName], props)
         app.use(createPinia());
         app.mount(appElement);
+    }
+
+    // mount globalnego alertu
+    const alertElement = document.getElementById('global-toast')
+
+    let alertAppInstance = null
+
+    if (alertElement) {
+        const alertApp = createApp(Toast)
+        alertAppInstance = alertApp.mount(alertElement)
+    }
+
+// globalna funkcja
+    window.$toast = (title, message, type = 'info', timeout = 3000) => {
+        if (alertAppInstance) {
+            debugger;
+            alertAppInstance.show(title, message, type, timeout)
+        }
     }
 }